Allow pretty much anything as a well-known name
The parser previously accepted a well-formed well-known name, or the
string "<none>". It's more useful to just accept anything, to let us
shoe-horn other stuff into the log format as needed (until there's a
better format); in particular, this will allow the field to be reused
for the process name in the short-term.