Align last of first-appearance headers to divider
The first time a bus connection is shown in the diagram, its names are
drawn, ideally aligned with the previous horizontal event divider.
However, previously the first of n bus names was aligned with the
divider, so the other ones spilled over into the diagram.

This has the interesting side effect that, if the first connection to be
mentioned has more than two names, only the last two are shown because
the rest are off the top of the diagram!