Don't leave bonus padding on the LHS in one-bus logs
This was because we were drawing a rule to just before where the first
system bus header would be, even if there is never going to be one.

The fix does mean a very slight rendering change: the first rule on a
one-bus log is ever-so-slightly shorter. But life is too short for me to
care. The rules are ugly anyway. :)