Fix keyboard scrolling of diagram
Previously we hooked into the window's key-press-event signal. This was
fine when there was only one widget... :)

But it turns out that if we make the layout focusable, and give it the
window focus, then we can take advantage of GtkScrolledWindow's built-in
[Ctrl-]PageUp/PageDown/Home/End support (as well as
Ctrl-Up/Dn/Left/Right) and only need to do our own thing for bare arrows
plus spacebar. There is currently no indication that the diagram has
focus, but nor is there in my web browser. :)