Canvas: invalidate entire width when moving highlight
Previously only the first bit of the highlight would be updated
properly, and the rest would only get redrawn when the window was
redrawn for some other reason. This is the crux of the reason why:

+    (width, _height) <- layoutGetSize layout
...
-        width <- liftM (fromIntegral . snd) (layoutGetSize layout)

But while we're here, move some duplicated code into invalidateRect, and
remove an unnecessary Int -> Double -> Int roundtrip.