Don't crash on UnmarshalError
We hit UnmarshalError if the log contains messages using types like 'h'
that dbus-core doesn't understand yet. We shouldn't crash when this
happens.

It would be better to show the warnings in the UI, so the user knows
some messages are missing; or maybe even interleave them so we can show
a warning inline to say “there should be a message here but we couldn't
read it”. Another time.

<https://bugs.freedesktop.org/show_bug.cgi?id=44714>