monitor: signal all messages up to the application
This is kind of a regression because we get all the internal guff too,
so the counter in the UI is wrong. But the loader needs all the
messages, even the internal ones, to track name changes.

We'll later update the recorder not to count messages that don't show
any output, and not to feed them to the renderer.