Renderer: fix confused thinking about edgemost apps
Previously, 'edgemostApp' threw in the coordinates for the next column as
well as for all the active columns, citing a poorly-specified rendering
glitch on the system bus logs. This turned out to specifically be that
one end of all signal lines was wildly misplaced. This in turn was due
to the per-bus sign stuff in 'signal' being upside-down: that function
was adding/removing one column width to compensate for the extra column
width added by edgemostApp (!), but in the wrong direction… so if we
just knocked both out we're grand.

In turn, the signs in getLeftMargin and getRightMargin (which add a
little padding so that the horizontal rules extend to roughly half a
column width past the outermost vertical line) were upside-down to
compensate for the extra column width being added.