Cope with gtk2hs's migration from String to Text
Most functions accept a GlibString instance, so for them we just need to
explicitly annotate the type as String.

Event names are now Text, but the pattern-matches in Canvas can be
fixed with OverloadedStrings.

GError's message is now a Text. I am lazy, so just 'show' it. It will be
equally ugly with old and new gtk2hs.