Rename Markup → Marquee
This:

    import Graphics.UI.Gtk hiding (Markup)

is a warning (which is apparently fatal in Travis CI-land) because new
gtk2hs doesn't export Markup (a synonym for String) from there. It's
only necessary because my own Markup shadowed pango's Markup, so let's
rename it.

Also, remove the 'show' calls: apparently Past Me added a typeclass to
coerce String or Text to String, so I can use that.