travis: install gtk2hs-buildtools and deps together
Installing these separately breaks with GHC 7.8.4 on Travis. First
gtk2hs-buildtools pulls in primitive, which gets built against
transformers-0.3.0.0. Then some dependency of Bustle pulls in
transformers-0.4.3.0 and wants to rebuild primitive against it. So my
hypothesis is that installing everything together will arrange for
everything to be built against the newer transformers and everyone will
be happy.