cabal: bump Cabal-Version
This version introduced support for setup-tools which is used by gtk et
al. to build-depend on gtk2hs-buildtools.

You can probably still build on older versions if you can arrange for
gtk2hs-buildtools to be installed but I couldn't get it to work, so here
we are.

While here: tweak Tested-With for multi-ghc-travis' benefit.