flatpak: build Haskell libs with -j$ncpus
flatpak-builder only exposes this parameter by passing it to `make`. I
guess in a better world I'd propose a patch to expose it in an
environment variable for build-commands, but here we are.

Rumour has it the FlatHub build workers have 8 (v)CPUs so I'd expect
this to make a big difference to the build time there.