CI: run flatpak-builder build, publish result
I combined the existing `stack build` and `stack test` build steps
because the logs show that the working directory is not shared between
them, so `stack test` re-runs the entire build (doubling the time for CI
to complete).

The Flatpak build takes an astonishing 30 minutes when uncached (not
sure why) but is very quick in the common case where all steps bar the
last one are cached.

https://gitlab.gnome.org/GNOME/Initiatives/wikis/DevOps-with-Flatpak and
https://docs.gitlab.com/ee/ci/yaml/README.html were both invaluable
resources.